FAC-647:feature:Merge UseCaseParam tests
